Initial-Commit
This project was built using JavaScript, React, React-Bootstrap, and Styled-Components.

The User may enter a movie title in the search bar and receive a collection of movies along the first row that match the search term.

Here there are two stages of API calls being done, the first is to get an array of results back, then a second calls are being done individually for each Title on screen so that we may populate the additional info on hover.

The user may hover over a movie poster to view more info on the film.

The user may then favorite the film by clicking the overlay on the bottom of the film when hovered. This data is saved locally in the browser data to restore favorites after coming back to the app.

This project was possible through The Open Movie Database.